实现recycler，未成功

+++ app/src/main/java/cn/ucai/fulicenter/bean/NewGoodBean.java
@@ -44,7 +44,7 @@ public class NewGoodBean {
     private String colorName;
     private String colorCode;
     private String colorUrl;
-    private int addTime;
+    private long addTime;
     private boolean promote;
 
     public int getId() {
@@ -208,11 +208,11 @@ public class NewGoodBean {
         this.colorUrl = colorUrl;
     }
 
-    public int getAddTime() {
+    public long getAddTime() {
         return addTime;
     }
 
-    public void setAddTime(int addTime) {
+    public void setAddTime(long addTime) {
         this.addTime = addTime;
     }
